Abstract out test functions.
